adding dateformater to a class
